+++ uCNC/src/hal/boards/avr/avr.ini
@@ -11,6 +11,7 @@ extra_scripts = avr_compiler.py
 ; debug_tool = simavr
 debug_build_flags = -Og -g3 -ggdb3 -gdwarf-2
 build_flags = ${common.build_flags} -mcall-prologues -mrelax -flto -fno-fat-lto-objects -fno-tree-scev-cprop -Wl,--relax
+lib_ignore = EEPROM, SPI, Wire
 
 [env:uno]
 extends = common_avr
