--- uCNC/src/hal/mcus/esp8266/esp8266_eeprom.cpp
@@ -1,34 +1,0 @@
-#include "../../../../cnc_config.h"
-#ifdef ESP8266
-#ifndef RAM_ONLY_SETTINGS
-#include <Arduino.h>
-#include <EEPROM.h>
-#include <stdint.h>
-extern "C"
-{
-	void esp8266_eeprom_init(int size)
-	{
-		EEPROM.begin(size);
-	}
-
-	uint8_t esp8266_eeprom_read(uint16_t address)
-	{
-		return EEPROM.read(address);
-	}
-
-	void esp8266_eeprom_write(uint16_t address, uint8_t value)
-	{
-		EEPROM.write(address, value);
-	}
-
-	void esp8266_eeprom_flush(void)
-	{
-		if (!EEPROM.commit())
-		{
-			Serial.println("[MSG: EEPROM write error]");
-		}
-	}
-}
-
-#endif
-#endif
