+++ uCNC/src/hal/mcus/esp8266/mcumap_esp8266.h
@@ -1053,8 +1053,6 @@ extern "C"
 		(uint8_t)(SPI1W0 & 0xff); \
 	}
 
-#define mcu_spi_config(X, Y) esp8266_spi_config(X, Y)
-
 	extern void esp8266_delay_us(uint16_t delay);
 #define mcu_delay_us(X) esp8266_delay_us(X)
