+++ uCNC/src/hal/mcus/lpc176x/mcu_lpc176x.c
@@ -584,7 +584,7 @@ void mcu_uart_putc(uint8_t c)
 
 void mcu_uart_flush(void)
 {
-	if (CHECKBIT(COM_UART->LSR, 5) && !(COM_UART->IER & UART_IER_THREINT_EN)) // not ready start flushing
+	if (!(COM_UART->IER & UART_IER_THREINT_EN)) // not ready start flushing
 	{
 		if (BUFFER_EMPTY(uart))
 		{
@@ -592,6 +592,8 @@ void mcu_uart_flush(void)
 		}
 		uint8_t c = 0;
 		BUFFER_DEQUEUE(uart, &c);
+		while (!CHECKBIT(COM_UART->LSR, 5))
+			;
 		COM_OUTREG = c;
 		COM_UART->IER |= UART_IER_THREINT_EN;
 #if ASSERT_PIN(ACTIVITY_LED)
@@ -613,7 +615,7 @@ void mcu_uart2_putc(uint8_t c)
 
 void mcu_uart2_flush(void)
 {
-	if (CHECKBIT(COM2_UART->LSR, 5) && !(COM2_UART->IER & UART_IER_THREINT_EN)) // not ready start flushing
+	if (!(COM2_UART->IER & UART_IER_THREINT_EN)) // not ready start flushing
 	{
 		if (BUFFER_EMPTY(uart2))
 		{
@@ -621,6 +623,8 @@ void mcu_uart2_flush(void)
 		}
 		uint8_t c = 0;
 		BUFFER_DEQUEUE(uart2, &c);
+		while (!CHECKBIT(COM2_UART->LSR, 5))
+			;
 		COM2_OUTREG = c;
 		COM2_UART->IER |= UART_IER_THREINT_EN;
 #if ASSERT_PIN(ACTIVITY_LED)
