+++ uCNC/src/hal/mcus/stm32f4x/mcu_stm32f4x.c
@@ -507,7 +507,7 @@ void mcu_uart_putc(uint8_t c)
 void mcu_uart_flush(void)
 {
 
-	if ((COM_UART->SR & USART_SR_TXE)) // not ready start flushing
+	if (!(COM_UART->CR1 & USART_CR1_TXEIE)) // not ready start flushing
 	{
 		COM_UART->CR1 |= (USART_CR1_TXEIE);
 #if ASSERT_PIN(ACTIVITY_LED)
@@ -531,7 +531,7 @@ void mcu_uart2_putc(uint8_t c)
 
 void mcu_uart2_flush(void)
 {
-	if ((COM2_UART->SR & USART_SR_TXE)) // not ready start flushing
+	if (!(COM2_UART->CR1 & USART_CR1_TXEIE)) // not ready start flushing
 	{
 		COM2_UART->CR1 |= (USART_CR1_TXEIE);
 #if ASSERT_PIN(ACTIVITY_LED)
